Moved TestRunner and Hooks to Helpers
