+++ src/test/java/helpers/Hooks.java
@@ -1,4 +1,4 @@
-package utility;
+package helpers;
 
 import java.net.MalformedURLException;
