+++ src/test/java/helpers/TestRunner.java
@@ -1,4 +1,4 @@
-package utility;
+package helpers;
 
 import org.junit.runner.RunWith;
 
@@ -7,7 +7,7 @@ import cucumber.api.junit.Cucumber;
 
 @RunWith(Cucumber.class)
 @CucumberOptions(features = "classpath:features", plugin = { "pretty",
-		"html:target/cucumber-html-report", "json:cucumber.json" }, tags = {},glue = {"utility","step_definitions"})
+		"html:target/cucumber-html-report", "json:cucumber.json" }, tags = {},glue = {"helpers","step_definitions"})
 public class TestRunner {
 
 }
